+++ flor/journal/entry/data/dataframe.py
@@ -1,0 +1,10 @@
+import json
+from pathlib import PurePath
+from typing import Union
+
+import pandas
+
+from flor import shelf
+
+from ..constants import *
+from .abstract import Data
