+++ flor/journal/entry/data/torch_chkpt.py
@@ -1,17 +1,16 @@
-import json
-from copy import deepcopy
 from pathlib import PurePath
 from typing import Union
 
-import cloudpickle
-
-from flor import shelf
+import json
 
 from ..constants import *
 from .abstract import Data
+from flor import shelf
+
+import torch
 
 
-class Reference(Data):
+class Torch(Data):
     def __init__(self, sk, gk, v=None, r: Union[None, PurePath] = None):
         assert bool(v is not None) != bool(r is not None)
         super().__init__(sk, gk, v)
@@ -20,14 +19,10 @@ class Reference(Data):
 
     def make_val(self):
         assert self.ref is not None
-        with open(self.ref, "rb") as f:
-            self.value = cloudpickle.load(f)
+        self.value = torch.load(self.ref)
 
     def would_mat(self):
-        """
-        For timing serialization costs
-        """
-        cloudpickle.dumps(self.value)
+        return
 
     def jsonify(self):
         assert self.ref is not None
@@ -46,8 +41,7 @@ class Reference(Data):
         assert (
             isinstance(self.ref, PurePath) and self.ref.suffix == PKL_SFX
         ), "Must first set a reference path with a `.pkl` suffix"
-        with open(self.ref, "wb") as f:
-            cloudpickle.dump(self.value, f)
+        torch.save(self.value, self.ref)
         self.val_saved = True
         self.value = None
 
@@ -67,12 +61,11 @@ class Reference(Data):
         )
 
     def promise(self):
-        self.promised = deepcopy(self.value)
-        self.value = self.promised
-
-    def fulfill(self):
-        super().fulfill()
         ref = shelf.get_pkl_ref()
         assert ref is not None
         self.set_ref_and_dump(ref)
-        return json.dumps(self.jsonify())
+        self.promised = json.dumps(self.jsonify())
+
+    def fulfill(self):
+        super().fulfill()
+        return self.promised
