+++ flor/logger/__init__.py
@@ -1,2 +1,3 @@
 from .logger import Logger
 from .future import Future
+from copy import deepcopy
